+++ src/Type/ProxyQueryDynamicReturnTypeExtension.php
@@ -19,6 +19,8 @@ use PHPStan\Reflection\BrokerAwareExtension;
 use PHPStan\Reflection\ClassReflection;
 use PHPStan\Reflection\MethodReflection;
 use PHPStan\Reflection\MethodsClassReflectionExtension;
+use Sonata\AdminBundle\Datagrid\ProxyQueryInterface as AdminProxyQueryInterface;
+use Sonata\DatagridBundle\ProxyQuery\ProxyQueryInterface as DatagridProxyQueryInterface;
 use Sonata\DoctrineORMAdminBundle\Datagrid\ProxyQuery;
 
 /**
@@ -44,8 +46,8 @@ class ProxyQueryDynamicReturnTypeExtension implements MethodsClassReflectionExte
      */
     public function hasMethod(ClassReflection $classReflection, string $methodName): bool
     {
-        return ProxyQuery::class === $classReflection->getName()
-            && $this->broker->getClass(QueryBuilder::class)->hasMethod($methodName);
+        return (\in_array($classReflection->getName(), [AdminProxyQueryInterface::class, DatagridProxyQueryInterface::class, ProxyQuery::class])
+            && $this->broker->getClass(QueryBuilder::class)->hasMethod($methodName));
     }
 
     /**