+++ CHANGELOG.md
@@ -8,3 +8,4 @@ master
 * Add coveralls and badges
 * Uprgade phpstan & phpstan-phpunit to ^0.11
 * Move phpstan-phpunit to dev dependencies
+* Improve detection of QueryBuilder methods